[#60] feat: SimpleMemberResponse 경로 이동 및 구현
